backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.0
docxtpl==0.20.2
Jinja2==3.1.6
kombu==5.6.2
lxml==6.0.2
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
redis>=5.0.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
dj-database-url
djangorestframework-simplejwt>=5.0.0
openai>=1.0.0
# OCR & AI
paddlepaddle
paddleocr
pymupdf
opencv-python-headless
diff-match-patch
numpy>=1.26.0
pandas>=2.1.0
openpyxl
gspread==6.2.1
google-auth>=2.20.0
google-generativeai>=0.8.0

